+++ app/src/main/java/com/mapbox/maps/testapp/examples/GesturesActivity.kt
@@ -259,7 +259,6 @@ class GesturesActivity : AppCompatActivity() {
   }
 
   private fun fixedFocalPointEnabled(enabled: Boolean) {
-
     if (enabled) {
       focalPointLatLng = FOCAL_POINT
       pointAnnotationManager =